backend/requirements.txt
asgiref==3.2.7
certifi==2019.11.28
chardet==3.0.4
Django==3.0.4
django-cors-headers==3.3.0
django-extensions==2.1.6
django-split-settings==1.0.0
django-python3-ldap==0.11.3
djangorestframework==3.11.0
idna==2.9
jupyter==1.0.0
ldap3==2.7
notebook==6.0.1
oauthlib==3.1.0
psycopg2==2.8.5
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.3.1
requests==2.23.0
requests-oauthlib==1.3.0
six==1.14.0
sqlparse==0.3.1
urllib3==1.25.8
xlrd==1.2.0
django-import-export==2.2.0
django-filter==2.2.0
